feat: stabilize trie-map exports, disable io_uring by default

+++ wrappers/src/basic_multi_key/mapx_double_key/mod.rs
@@ -19,7 +19,7 @@ use std::{
     ops::{Deref, DerefMut},
 };
 
-const KEY_SIZE: usize = 2;
+const KEY_SIZE: u32 = 2;
 
 /// A map structure with two-level keys.
 #[derive(Serialize, Deserialize, Debug)]
@@ -116,7 +116,7 @@ where
     }
 
     #[inline(always)]
-    pub fn key_size(&self) -> usize {
+    pub fn key_size(&self) -> u32 {
         self.inner.key_size()
     }
